gpo_cloud_dashboard/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
WTForms==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Faker==20.1.0
email-validator==2.1.0.post1
pytest==7.4.3
pytest-flask==1.3.0 